Island: Split up utils in repository
